+++ modules/seminar/module.mk
@@ -24,7 +24,7 @@ define prepare_arguments_round
 endef
 
 build/seminar/%/intro.tex build/seminar/%/rules.tex: \
-	modules/seminar/templates/$$(notdir $@)
+	modules/seminar/templates/$$(notdir $$@)
 	$(call _prepare_arguments)
 	python -m modules.seminar.builder.volume 'source/seminar/' 'source/seminar/$*/' \
 		-c $(word 1,$(words)) -v $(word 2,$(words)) -o '$(dir $@)' || exit 1;
@@ -39,12 +39,12 @@ build/seminar/%/semester.tex: \
 	$(call prepare_arguments_semester,semester)
 
 build/seminar/%/invite.tex: \
-	modules/seminar/templates/$$(notdir $@) \
+	modules/seminar/templates/$$(notdir $$@) \
 	source/seminar/$$*/meta.yaml
 	$(call prepare_arguments_semester,invite)
 
 build/seminar/%/problems.tex build/seminar/%/solutions.tex build/seminar/%/solutions-full.tex build/seminar/%/instagram.tex: \
-	modules/seminar/templates/$$(notdir $@) \
+	modules/seminar/templates/$$(notdir $$@) \
 	$$(wildcard source/seminar/$$*/*/meta.yaml) \
 	source/seminar/$$*/meta.yaml
 	$(call prepare_arguments_round,round)
@@ -125,7 +125,7 @@ output/seminar/%/instagram: \
 output/seminar/%/html-problems: \
 	output/seminar/$$*/html-prerequisites \
 	$$(subst source/,output/,$$(subst .md,.html,$$(wildcard source/seminar/$$*/*/problem.md))) ;
-	
+
 output/seminar/%/html-solutions:\
 	output/seminar/$$*/html-prerequisites \
 	$$(subst source/,output/,$$(subst .md,.html,$$(wildcard source/seminar/$$*/*/solution.md))) ;
